Updated - cleaned code a little - all regularPiece killings validated - end game after 12 kills implemented - started uberPiece implementation (only advances, no movement updated)
