--- src/main/java/com/sda/games/checkers/model/RegularPiece.java
@@ -1,9 +1,0 @@
-package com.sda.games.checkers.model;
-
-public class RegularPiece extends Piece{
-
-    public RegularPiece(boolean white) {
-        super(white);
-    }
-
-}
