+++ src/main/java/com/sda/games/checkers/model/game/GameStatus.java
@@ -1,4 +1,4 @@
-package com.sda.games.checkers.model;
+package com.sda.games.checkers.model.game;
 
 public enum GameStatus {
     ACTIVE,
