+++ src/main/java/com/sda/games/checkers/model/piece/Piece.java
@@ -1,0 +1,24 @@
+package com.sda.games.checkers.model.piece;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public abstract class Piece {
+
+    public boolean white;
+    private boolean regular;
+    private boolean killed;
+    private String pieceIcon;
+
+    public String getPieceIcon() {
+        return this.pieceIcon;
+    }
+
+    public Piece(boolean white) {
+        this.white = white;
+    }
+
+}
